+++ src/model/ccButton.h
@@ -23,7 +23,7 @@ public:
              bool needsRead = false);
     virtual Led getLed(byte value) override;
     virtual bool changed() override;
-    virtual void update(byte value) override;
+    virtual void update(int value) override;
 };
 
 #endif
