+++ src/model/effectButton.h
@@ -18,7 +18,7 @@ public:
                   unsigned long typeParameter);
     virtual Led getLed(byte value) override;
     virtual bool changed() override;
-    virtual void update(byte value) override;
+    virtual void update(int value) override;
     byte getState();
     byte getType();
     unsigned long getTypeParameter();
