+++ src/model/tapButton.h
@@ -17,7 +17,7 @@ public:
               unsigned long readParameter);
     virtual Led getLed(byte value) override;
     virtual bool changed() override;
-    virtual void update(byte value) override;
+    virtual void update(int value) override;
 };
 
 #endif
